Generalized the hyperlink rules.
