+++ automobile/appendix.md
@@ -4,5 +4,5 @@
 - [C++ libraries](cpp-libraries.md)
 - [Java libraries](java-libraries.md)
 - [Python libraries](python-libraries.md)
-- [ros_automobile controller](ros_automobile-controller.md)
+- [ros\_automobile controller](ros_automobile-controller.md)
 - [A typical OpenstreetMap importer configuration file](a-typical-openstreetmap-importer-configuration-file.md)
