+++ automobile/ros_automobile-controller.md
@@ -32,7 +32,7 @@ available with the `ros_automobile` controller.
 | [/automobile/get\_indicator](driver-library.md#wbu_driver_set_dipped_beams)                                                                                                                                                                                                                                                                                      | service       | webots\_ros::get\_bool                                       |                                                                                                                                               |
 | [/automobile/get\_indicator\_period](car-library.md#wbu_car_set_indicator_period)                                                                                                                                                                                                                                                                                | service       | webots\_ros::get\_float                                      |                                                                                                                                               |
 | [/automobile/get\_type](car-library.md#wbu_car_get_type)                                                                                                                                                                                                                                                                                                         | service       | webots\_ros::get\_int                                        |                                                                                                                                               |
-| [/automobile/get\_wipers_mode](driver-library.md#wbu_driver_set_wipers_mode)                                                                                                                                                                                                                                                                                      | service       | webots\_ros::get\_int                                       |                                                                                                                                               |
+| [/automobile/get\_wipers\_mode](driver-library.md#wbu_driver_set_wipers_mode)                                                                                                                                                                                                                                                                                     | service       | webots\_ros::get\_int                                       |                                                                                                                                               |
 | [/automobile/set\_antifog\_light](driver-library.md#wbu_driver_set_dipped_beams)                                                                                                                                                                                                                                                                                 | service       | webots\_ros::set\_bool                                       |                                                                                                                                               |
 | [/automobile/set\_brake\_intensity](driver-library.md#wbu_driver_set_brake_intensity)                                                                                                                                                                                                                                                                            | service       | webots\_ros::set\_float                                      |                                                                                                                                               |
 | [/automobile/set\_cruising\_speed](driver-library.md#wbu_driver_set_cruising_speed)                                                                                                                                                                                                                                                                              | service       | webots\_ros::set\_float                                      |                                                                                                                                               |
